# Stream query responses from boltdb index client