--- a/pkg/storage/chunk/local/boltdb_index_client.go
+++ b/pkg/storage/chunk/local/boltdb_index_client.go
@@ -268,7 +268,15 @@ func (b *BoltIndexClient) QueryWithCursor(_ context.Context, c *bbolt.Cursor, qu
 
 	rowPrefix := []byte(query.HashValue + separator)
 
-	var batch boltReadBatch
+	// sync.WaitGroup is needed to wait for the caller to finish processing all the index entries being streamed
+	wg := sync.WaitGroup{}
+	batch := newReadBatch()
+	defer func() {
+		batch.done()
+		wg.Wait()
+	}()
+
+	callbackDone := false
 
 	for k, v := c.Seek(start); k != nil; k, v = c.Next() {
 		if !bytes.HasPrefix(k, rowPrefix) {
@@ -282,16 +290,33 @@ func (b *BoltIndexClient) QueryWithCursor(_ context.Context, c *bbolt.Cursor, qu
 			continue
 		}
 
+		// we need to do callback only once to pass the batch iterator
+		if !callbackDone {
+			wg.Add(1)
+			// do the callback in a goroutine to stream back the index entries
+			go func() {
+				// wait for callback to finish processing the batch and return
+				defer wg.Done()
+				callback(query, batch)
+			}()
+			callbackDone = true
+		}
+
 		// make a copy since k, v are only valid for the life of the transaction.
 		// See: https://godoc.org/github.com/boltdb/bolt#Cursor.Seek
-		batch.rangeValue = make([]byte, len(k)-len(rowPrefix))
-		copy(batch.rangeValue, k[len(rowPrefix):])
+		rangeValue := make([]byte, len(k)-len(rowPrefix))
+		copy(rangeValue, k[len(rowPrefix):])
 
-		batch.value = make([]byte, len(v))
-		copy(batch.value, v)
+		value := make([]byte, len(v))
+		copy(value, v)
 
-		if !callback(query, &batch) {
-			break
+		err := batch.send(singleResponse{
+			rangeValue: rangeValue,
+			value:      value,
+		})
+
+		if err != nil {
+			return err
 		}
 	}
 
@@ -334,36 +359,49 @@ func (b *BoltWriteBatch) Add(tableName, hashValue string, rangeValue []byte, val
 	writes.puts[key] = value
 }
 
-type boltReadBatch struct {
+type singleResponse struct {
 	rangeValue []byte
 	value      []byte
 }
 
-func (b boltReadBatch) Iterator() chunk.ReadBatchIterator {
-	return &boltReadBatchIterator{
-		boltReadBatch: b,
+type readBatch struct {
+	respChan chan singleResponse
+	curr     singleResponse
+}
+
+func newReadBatch() *readBatch {
+	return &readBatch{respChan: make(chan singleResponse)}
+}
+
+func (r *readBatch) Iterator() chunk.ReadBatchIterator {
+	return r
+}
+
+func (r *readBatch) Next() bool {
+	var ok bool
+	r.curr, ok = <-r.respChan
+	return ok
+}
+
+func (r *readBatch) RangeValue() []byte {
+	return r.curr.rangeValue
+}
+
+func (r *readBatch) Value() []byte {
+	return r.curr.value
+}
+
+func (r *readBatch) done() {
+	close(r.respChan)
+}
+
+func (r *readBatch) send(resp singleResponse) error {
+	select {
+	case r.respChan <- resp:
+		return nil
+	case <-time.After(10 * time.Second):
+		return errors.New("timed out sending response")
 	}
-}
-
-type boltReadBatchIterator struct {
-	consumed bool
-	boltReadBatch
-}
-
-func (b *boltReadBatchIterator) Next() bool {
-	if b.consumed {
-		return false
-	}
-	b.consumed = true
-	return true
-}
-
-func (b *boltReadBatchIterator) RangeValue() []byte {
-	return b.rangeValue
-}
-
-func (b *boltReadBatchIterator) Value() []byte {
-	return b.value
 }
 
 // Open the database.