--- a/pkg/storage/stores/shipper/indexgateway/gateway_test.go
+++ b/pkg/storage/stores/shipper/indexgateway/gateway_test.go
@@ -128,7 +128,7 @@ func TestGateway_QueryIndex(t *testing.T) {
 		}
 		expectedQueryKey = util.QueryKey(query)
 
-		gateway.shipper = mockIndexClient{response: &mockBatch{size: responseSize}}
+		gateway.indexQuerier = mockIndexClient{response: &mockBatch{size: responseSize}}
 		err := gateway.QueryIndex(&indexgatewaypb.QueryIndexRequest{Queries: []*indexgatewaypb.IndexQuery{{
 			TableName:        query.TableName,
 			HashValue:        query.HashValue,