**What this PR does / why we need it**:
Previously, we sent just one row per batch in the index query response callback from the boltdb index client.
This, in turn, makes us send just one row at a time per grpc response from the index gateway server.

This PR improves it by streaming the response from the boltdb index client, which would, in turn, lets us send back a larger batch in grpc response from the index gateway server.

Here is the benchmark comparing the performance difference:
```
benchmark                                     old ns/op      new ns/op      delta
Benchmark_QueriesMatchingLargeNumOfRows-8     2027105841     1298895890     -35.92%

benchmark                                     old allocs     new allocs     delta
Benchmark_QueriesMatchingLargeNumOfRows-8     29321207       10141758       -65.41%

benchmark                                     old bytes      new bytes     delta
Benchmark_QueriesMatchingLargeNumOfRows-8     1032313592     507791808     -50.81%
```

This PR also adds the benchmarks to the repository.

**Checklist**
- [x] Tests updated